Changed Gene to double

+++ GeneticAlgorithm/Chromosome.cs
@@ -2,13 +2,13 @@
 {
     public class Chromosome
     {
-        public Chromosome(int[] genes)
+        public Chromosome(double[] genes)
         {
             Genes = genes;
         }
 
         public static int Length { get; set; }
-        public int[] Genes { get; set; }
+        public double[] Genes { get; set; }
         public double Fitness { get; set; }
 
         public override string ToString()
