+++ MathEquations/Equation1.cs
@@ -10,7 +10,7 @@ namespace MathEquations
 
         public const int VariableCount = 4;
 
-        public static double CalculateFitness(IEnumerable<int> variables)
+        public static double CalculateFitness(IEnumerable<double> variables)
         {
             var eqVar = variables.ToArray();
             var fitness = Math.Abs(eqVar[0] + 2 * eqVar[1] + 3 * eqVar[2] + 4 * eqVar[3] - 30);
